pyproject.toml
[tool.poetry]
name = "labforge-api"
version = "0.1.0"
description = "A FastAPI backend API for managing labs/training for UI"
authors = ["DevOps-With-Brian"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pydantic = {version = "^2.10.2", extras = ["email"]}
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
aiosqlite = "^0.20.0"
alembic = "^1.13.3"
greenlet = "^3.1.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.1"
ruff = "^0.8.1"
psycopg = {version = "^3.2.3", extras = ["binary"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
